# Updated rendering strategy and progress bars