--- a/src/FirstOrderMethods/callbacks.jl
+++ b/src/FirstOrderMethods/callbacks.jl
@@ -1,6 +1,6 @@
-function ensure_domain(m::AbstractFirstOrderMetricParams{T}) where {T}
+function ensure_domain(m::AbstractFirstOrderMetricParams{T}, closest_approach, effective_infinity) where {T}
     min_r = inner_radius(m)
-    (u, λ, integrator) -> u[2] ≤ min_r * 1.1 || u[2] > 1200.0
+    (u, λ, integrator) -> u[2] ≤ min_r * closest_approach || u[2] > effective_infinity
 end
 
 function flip_radial_sign!(integrator)