--- a/src/FirstOrderMethods/callbacks.jl
+++ b/src/FirstOrderMethods/callbacks.jl
@@ -1,4 +1,8 @@
-function ensure_domain(m::AbstractFirstOrderMetricParams{T}, closest_approach, effective_infinity) where {T}
+function ensure_domain(
+    m::AbstractFirstOrderMetricParams{T},
+    closest_approach,
+    effective_infinity,
+) where {T}
     min_r = inner_radius(m)
     (u, λ, integrator) -> u[2] ≤ min_r * closest_approach || u[2] > effective_infinity
 end