--- a/src/GeodesicTracer/GeodesicTracer.jl
+++ b/src/GeodesicTracer/GeodesicTracer.jl
@@ -51,18 +51,22 @@ function tracegeodesics(
     time_domain::Tuple{T,T};
     solver = Tsit5(),
     μ = 0.0,
+    closest_approach=1.01,
+    effective_infinity=1200.0,
     solver_opts...,
 ) where {T}
     __tracegeodesics(
         m,
         init_positions,
         # ensure everything already normalised
-        constrain_all(m, init_positions, init_velocities, T(μ)),
+        constrain_all(m, init_positions, init_velocities, μ),
         time_domain,
         solver;
         callback = nothing,
         abstol = 1e-9,
         reltol = 1e-9,
+        closest_approach = closest_approach,
+        effective_infinity = effective_infinity,
         solver_opts...,
     )
 end