--- a/src/GeodesicTracer/constraints.jl
+++ b/src/GeodesicTracer/constraints.jl
@@ -1,28 +1,3 @@
-# inplace
-function constrain_all(
-    m::AbstractMetricParams{T},
-    us::AbstractArray{T,2},
-    vs::AbstractArray{T,2},
-    μ,
-) where {T<:Number}
-    us_cols = eachcol(us)
-    vs_cols = eachcol(vs)
-    curried(u, v) = constrain(m, u, v, μ = μ)
-    @. vs[1, :] = curried(us_cols, vs_cols)
-    vs
-end
-
-# SMatrix -- do we actually want to support this?
-function constrain_all(
-    m::AbstractMetricParams{T},
-    us::SMatrix{4,M,T,L},
-    vs::SMatrix{4,M,T,L},
-    μ,
-) where {M,T<:Number,L}
-    alloc_vs = constrain_all(m, us, similar(vs), μ)
-    SMatrix{4,M}(alloc_vs)
-end
-
 # vectors of vectors 
 function constrain_all(m::AbstractMetricParams{T}, us, vs, μ) where {T}
     @inbounds for i in eachindex(vs)