--- a/src/GeodesicTracer/constraints.jl
+++ b/src/GeodesicTracer/constraints.jl
@@ -31,6 +31,11 @@ function constrain_all(
     @inbounds SVector{S,T}(constrain(m, u, v, μ = μ), v[2], v[3], v[4])
 end
 
-function wrap_constraint(m::AbstractMetricParams{T}, position, velfunc::Function, μ) where {T}
+function wrap_constraint(
+    m::AbstractMetricParams{T},
+    position,
+    velfunc::Function,
+    μ,
+) where {T}
     (i) -> constrain_all(m, position, velfunc(i), μ)
 end