--- a/src/GeodesicTracer/tracer.jl
+++ b/src/GeodesicTracer/tracer.jl
@@ -6,10 +6,12 @@ function __tracegeodesics(
     time_domain::Tuple{T,T},
     solver;
     callback,
+    closest_approach,
+    effective_infinity,
     kwargs...,
-) where {T}
+) where {T<:Number}
     prob = integrator_problem(m, init_pos, init_vel, time_domain)
-    cbs = create_callback_set(m, callback)
+    cbs = create_callback_set(m, callback, closest_approach, effective_infinity)
     solve_geodesic(solver, prob; callback = cbs, kwargs...)
 end
 
@@ -22,6 +24,8 @@ function __tracegeodesics(
     solver;
     ensemble = EnsembleThreads(),
     callback,
+    closest_approach, 
+    effective_infinity,
     kwargs...,
 ) where {T}
     prob = integrator_problem(
@@ -41,7 +45,7 @@ function __tracegeodesics(
         safetycopy = false,
     )
 
-    cbs = create_callback_set(m, callback)
+    cbs = create_callback_set(m, callback, closest_approach, effective_infinity)
     solve_geodesic(
         solver,
         ens_prob,
@@ -61,6 +65,8 @@ function __tracegeodesics(
     solver;
     ensemble = EnsembleThreads(),
     callback,
+    closest_approach, 
+    effective_infinity,
     kwargs...,
 ) where {T}
     prob = integrator_problem(m, init_positions[1], init_velocities[1], time_domain)
@@ -71,7 +77,7 @@ function __tracegeodesics(
         safetycopy = false,
     )
 
-    cbs = create_callback_set(m, callback)
+    cbs = create_callback_set(m, callback, closest_approach, effective_infinity)
     solve_geodesic(
         solver,
         ens_prob,