--- a/src/GeodesicTracer/tracer.jl
+++ b/src/GeodesicTracer/tracer.jl
@@ -15,11 +15,12 @@ function __tracegeodesics(
     solve_geodesic(solver, prob; callback = cbs, kwargs...)
 end
 
-# columnar of positions and velocity
+# single position and single velocity
+# ensembled with an indexable function
 function __tracegeodesics(
     m::AbstractMetricParams{T},
-    init_positions::AbstractArray{T,2},
-    init_velocities::AbstractArray{T,2},
+    init_pos::AbstractVector{T},
+    init_vel::AbstractVector{T},
     time_domain::Tuple{T,T},
     solver;
     ensemble = EnsembleThreads(),
@@ -27,33 +28,10 @@ function __tracegeodesics(
     closest_approach, 
     effective_infinity,
     kwargs...,
-) where {T}
-    prob = integrator_problem(
-        m,
-        @view(init_positions[:, 1]),
-        @view(init_velocities[:, 1]),
-        time_domain,
-    )
-    ens_prob = EnsembleProblem(
-        prob,
-        prob_func = (prob, i, repeat) -> integrator_problem(
-            m,
-            @view(init_positions[:, i]),
-            @view(init_velocities[:, i]),
-            time_domain,
-        ),
-        safetycopy = false,
-    )
-
+) where {T<:Number}
+    prob = integrator_problem(m, init_pos, init_vel, time_domain)
     cbs = create_callback_set(m, callback, closest_approach, effective_infinity)
-    solve_geodesic(
-        solver,
-        ens_prob,
-        ensemble;
-        trajectories = size(init_velocities, 2),
-        callback = cbs,
-        kwargs...,
-    )
+    solve_geodesic(solver, prob; callback = cbs, kwargs...)
 end
 
 # indexables