--- a/src/GeodesicTracer/tracer.jl
+++ b/src/GeodesicTracer/tracer.jl
@@ -5,14 +5,10 @@ function __tracegeodesics(
     init_vel::AbstractVector{T},
     time_domain::Tuple{T,T},
     solver;
-    callback,
-    closest_approach,
-    effective_infinity,
     kwargs...,
 ) where {T<:Number}
     prob = integrator_problem(m, init_pos, init_vel, time_domain)
-    cbs = create_callback_set(m, callback, closest_approach, effective_infinity)
-    solve_geodesic(solver, prob; callback = cbs, kwargs...)
+    solve_geodesic(solver, prob; kwargs...)
 end
 
 # single position and single velocity
@@ -20,18 +16,28 @@ end
 function __tracegeodesics(
     m::AbstractMetricParams{T},
     init_pos::AbstractVector{T},
-    init_vel::AbstractVector{T},
+    vel_func::Function,
     time_domain::Tuple{T,T},
-    solver;
+    solver
+    ;
+    trajectories::Int,
     ensemble = EnsembleThreads(),
-    callback,
-    closest_approach, 
-    effective_infinity,
     kwargs...,
 ) where {T<:Number}
-    prob = integrator_problem(m, init_pos, init_vel, time_domain)
-    cbs = create_callback_set(m, callback, closest_approach, effective_infinity)
-    solve_geodesic(solver, prob; callback = cbs, kwargs...)
+    prob = integrator_problem(m, init_pos, vel_func(1), time_domain)
+    ens_prob = EnsembleProblem(
+        prob,
+        prob_func = (prob, i, repeat) ->
+            integrator_problem(m, init_pos, vel_func(i), time_domain),
+        safetycopy = false,
+    )
+    solve_geodesic(
+        solver,
+        ens_prob,
+        ensemble;
+        trajectories = trajectories,
+        kwargs...,
+    )
 end
 
 # indexables
@@ -42,9 +48,6 @@ function __tracegeodesics(
     time_domain::Tuple{T,T},
     solver;
     ensemble = EnsembleThreads(),
-    callback,
-    closest_approach, 
-    effective_infinity,
     kwargs...,
 ) where {T}
     prob = integrator_problem(m, init_positions[1], init_velocities[1], time_domain)
@@ -55,13 +58,11 @@ function __tracegeodesics(
         safetycopy = false,
     )
 
-    cbs = create_callback_set(m, callback, closest_approach, effective_infinity)
     solve_geodesic(
         solver,
         ens_prob,
         ensemble;
         trajectories = length(init_velocities),
-        callback = cbs,
         kwargs...,
     )
 end