I've updated the rendering strategy to avoid pre-allocating the same position and velocity vectors over and over, and instead calculate them on the fly (since they are deterministic). Downside is, we can no longer count which row we are on, since that's racy with multiple threads...

So instead, added some progress meters:
```
+ Starting trace...
Rendering: 100%[========================================] Time: 0:00:41
+ Trace complete.
 42.084828 seconds (1.48 G allocations: 51.313 GiB, 37.36% gc time, 1.00% compilation time)
```

Performance hit of this is negligible, but they can be disabled by passing `verbose = false` to `rendergeodesics`.

Test setup:
```julia
using Gradus, StaticArrays
m = BoyerLindquistAD(M=1.0, a=0.998)

u = @SVector [0.0, 1000.0, deg2rad(90), 0.0]

img = @time rendergeodesics(
    m, u, 2000.0; 
    abstol=1e-9, reltol=1e-9,
    image_width=200 * 4, 
    image_height=200 * 4, 
    fov_factor=13.0 * 4,
    verbose = true
)
```

Before, running on 12 threads:
```julia
# 55.465356 seconds (1.47 G allocations: 51.276 GiB, 33.66% gc time, 4.16% compilation time)
```
Now:
```julia
# 42.259254 seconds (1.29 G allocations: 45.719 GiB, 31.97% gc time, 0.04% compilation time)
```
Shaved 13 seconds off, which is a start.